requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
asgiref==3.7.2 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
boto3==1.28.1
botocore==1.31.1 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
cffi==1.15.1
channels==4.0.0 ; python_version >= '3.7'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-rest-auth==4.0.1
django==4.0.4
django-allauth==0.54.0
django-cors-headers==3.13.0
django-environ==0.8.1
django-nextjs==2.3.0
django-rest-auth==0.9.5
django-softdelete==0.10.5
djangorestframework==3.13.1
djangorestframework--simplejwt==5.2.2
drf-spectacular==0.26.3
drf-yasg==1.20.0
frozenlist==1.3.3 ; python_version >= '3.7'
gunicorn==20.1.0
idna==3.4 ; python_version >= '3.5'
inflection==0.5.1 ; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.2 ; python_version >= '3.7'
jmespath==1.0.1 ; python_version >= '3.7'
jsonschema==4.18.0 ; python_version >= '3.8'
jsonschema-specifications==2023.6.1 ; python_version >= '3.8'
markupsafe==2.1.3 ; python_version >= '3.7'
multidict==6.0.4 ; python_version >= '3.7'
oauthlib==3.2.2 ; python_version >= '3.6'
packaging==23.1 ; python_version >= '3.7'
pillow==9.2.0
psycopg2-binary==2.9.5
pycparser==2.21
pyjwt[crypto]==2.7.0 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python3-openid==3.2.0
pytz==2023.3
pyyaml==6.0 ; python_version >= '3.6'
referencing==0.29.1 ; python_version >= '3.8'
requests==2.31.0 ; python_version >= '3.7'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rpds-py==0.8.10 ; python_version >= '3.8'
ruamel.yaml==0.17.32 ; python_version >= '3'
ruamel.yaml.clib==0.2.7 ; python_version < '3.12' and platform_python_implementation == 'CPython'
s3transfer==0.6.1 ; python_version >= '3.7'
setuptools==68.0.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4 ; python_version >= '3.5'
typing-extensions==4.7.1 ; python_version < '3.11'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.16 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uwsgi==2.0.21
websockets==11.0.3 ; python_version >= '3.7'
wheel==0.40.0 ; python_version >= '3.7'
yarl==1.9.2 ; python_version >= '3.7'
